# perf: convert to string once in setCacheValue