@odeke-em 2021-08-24
Nice work with the improvement, it shows an improvement overall in other related code https://dashboard.github.orijtech.com/benchmark/2b046b0768ad4ef9b4b0c28163e7f112